Enable to use small-modules-for
