--- src/main/scala/org/scalajs/cli/internal/ModuleSplitStyleParser110Plus.scala
@@ -1,16 +1,0 @@
-package org.scalajs.cli.internal
-
-import org.scalajs.linker.interface.{ModuleSplitStyle => ActualModuleSplitStyle}
-
-// class rather than object, as that's easier to substitute from native-image
-class ModuleSplitStyleParser110Plus {
-  def parse(splitStyle: String, modulePackages: Array[String]): ModuleSplitStyle =
-    (new ModuleSplitStyleParser).tryParse(splitStyle).getOrElse {
-      if (splitStyle == "SmallModulesFor") {
-        if (modulePackages.isEmpty)
-          throw new IllegalArgumentException(s"SmallModuleFor style must have at least one package. To define it pass `--smallModuleForPackages` parameter.")
-        ModuleSplitStyle(ActualModuleSplitStyle.SmallModulesFor(modulePackages.toList))
-      } else
-        throw new IllegalArgumentException(s"$splitStyle is not a valid module split style")
-    }
-}
